Rename projection to place_ty
